weather_api/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-decouple==3.7
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.3
